requirements.txt
async-lru
autopep8
pyaes
pyasn1
pycodestyle
Pyrogram
PySocks
rsa
TgCrypto
requests
loguru
schedule
arrow
python-dotenv
emoji
requests_futures
argparse
pymongo
pytz
r7insight_python
dill
urlextract
youtube_dl
uuid
ffmpeg-python
youtube_search
pymongo
redis
dnspython
pillow
aiofiles
aiohttp
schedule
pytgcalls